+++ src/main/kotlin/y2019/Day10.kt
@@ -30,11 +30,11 @@ class Day10 : Day(2019, 10) {
                 if (otherAsteroid != asteroid) {
                     val dx = otherAsteroid.x - asteroid.x
                     val dy = otherAsteroid.y - asteroid.y
-                    if (dx == 0 && otherAsteroid == input.firstAsteroid(asteroid, 0, dy / abs(dy))) {
+                    if (dx == 0L && otherAsteroid == input.firstAsteroid(asteroid, 0, dy / abs(dy))) {
                         count++
-                    } else if (dy == 0 && otherAsteroid == input.firstAsteroid(asteroid, dx / abs(dx), 0)) {
+                    } else if (dy == 0L && otherAsteroid == input.firstAsteroid(asteroid, dx / abs(dx), 0)) {
                         count++
-                    } else if (dx != 0 && dy != 0) {
+                    } else if (dx != 0L && dy != 0L) {
                         val gcd = dx.gcd(dy)
                         if (otherAsteroid == input.firstAsteroid(asteroid, dx / gcd, dy / gcd)) {
                             count++
@@ -50,7 +50,7 @@ class Day10 : Day(2019, 10) {
         return maxAsteroids
     }
 
-    override fun solve2(): Int {
+    override fun solve2(): Long {
         val distances = mutableListOf<Point>()
         asteroids.filter { it != station }.forEach {
             distances += Point(station.x - it.x, station.y - it.y)
@@ -60,7 +60,7 @@ class Day10 : Day(2019, 10) {
         val map = distances.reversed().groupBy { atan2(it.y.toFloat(), it.x.toFloat()) }.toMutableMap()
 
         val angles = map.keys.toList().sorted().toMutableList()
-        val start = map.keys.first { it -> map[it]!!.any { it.x == 0 } }
+        val start = map.keys.first { it -> map[it]!!.any { it.x == 0L } }
         var index = angles.indexOf(start)
         var counter = 0
 
@@ -83,11 +83,11 @@ class Day10 : Day(2019, 10) {
         return 0
     }
 
-    private fun Array<CharArray>.firstAsteroid(start: Point, dx: Int, dy: Int): Point? {
+    private fun Array<CharArray>.firstAsteroid(start: Point, dx: Long, dy: Long): Point? {
         var x = start.x + dx
         var y = start.y + dy
-        while (x in indices && y in get(x).indices) {
-            if (get(x)[y] == '#') return Point(x, y)
+        while (x in indices && y in get(x.toInt()).indices) {
+            if (get(x.toInt())[y.toInt()] == '#') return Point(x, y)
             x += dx
             y += dy
         }
